moved joint ot mappings to the mapping module
