+++ ot/__init__.py
@@ -52,7 +52,8 @@ from .gromov import (gromov_wasserstein, gromov_wasserstein2,
 from .weak import weak_optimal_transport
 from .factored import factored_optimal_transport
 from .solvers import solve
-from .mapping import nearest_brenier_potential_fit, nearest_brenier_potential_predict_bounds
+from .mapping import (nearest_brenier_potential_fit, nearest_brenier_potential_predict_bounds, joint_OT_mapping_kernel,
+                      joint_OT_mapping_linear)
 
 # utils functions
 from .utils import dist, unif, tic, toc, toq
@@ -71,4 +72,4 @@ __all__ = ['emd', 'emd2', 'emd_1d', 'sinkhorn', 'sinkhorn2', 'utils',
            'smooth', 'stochastic', 'unbalanced', 'partial', 'regpath', 'solvers',
            'binary_search_circle', 'wasserstein_circle',
            'semidiscrete_wasserstein2_unif_circle', 'sliced_wasserstein_sphere_unif', 'nearest_brenier_potential_fit',
-           'nearest_brenier_potential_predict_bounds']
+           'nearest_brenier_potential_predict_bounds', 'joint_OT_mapping_kernel', 'joint_OT_mapping_linear']
